Txn.Epoch: int32->uint32

+++ storage/replica.go
@@ -1097,7 +1097,7 @@ func (r *Replica) applyRaftCommandInBatch(ctx context.Context, index uint64, ori
 		}
 		// Only transactional requests have replay protection.
 		if ba.Txn != nil {
-			if putErr := r.sequence.PutSequence(btch, ba.Txn.ID, uint32(ba.Txn.Epoch),
+			if putErr := r.sequence.PutSequence(btch, ba.Txn.ID, ba.Txn.Epoch,
 				ba.Txn.Sequence, ba.Txn.Key, ba.Txn.Timestamp, err); putErr != nil {
 				// TODO(tschottdorf): ReplicaCorruptionError.
 				log.Fatalc(ctx, "putting a sequence cache entry in a batch should never fail: %s", putErr)
